Flag now ui32.

+++ libmemseries/meas.h
@@ -12,7 +12,7 @@
 namespace memseries {
 	typedef uint64_t Time;
 	typedef uint64_t Id;
-	typedef uint64_t Flag;
+	typedef uint32_t Flag;
 	typedef double Value;
 	typedef std::vector<Id> IdArray;
 	typedef std::set<Id> IdSet;
